Node name added as a option for host display name
The host_display_name_attr is restricted to only fqdn and hostname
attributes of a node, only these could be used to name a host in icinga2.
This commit adds another option - a node name, under which is a node
registered within chef server.
For me this is useful, because nor fqdn or hostname are not really
descriptive and can't/won't change it. But I can use the chef node name,
which is under my control, nice and descriptive.