Be more conservative with Chef attributes
While '-' literals are valid in attribute keys, they are not valid with
'dot' notation.

Plug in icinga_format so that differences between Ruby and the Icinga
configuration language can be papered over comprehensively.

This is the only template I needed to touch for my use case, but there
are others that seem like they could use the same treatment.